+++ src/posthog-core.ts
@@ -46,7 +46,7 @@ import {
     SnippetArrayItem,
     ToolbarParams,
 } from './types'
-import { SentryIntegration } from './extensions/sentry-integration'
+import { SentryIntegration, SentryIntegrationOptions, sentryIntegration } from './extensions/sentry-integration'
 import { setupSegmentIntegration } from './extensions/segment-integration'
 import { PageViewManager } from './page-view'
 import { PostHogSurveys } from './posthog-surveys'
@@ -72,6 +72,7 @@ import { Heatmaps } from './heatmaps'
 import { ScrollManager } from './scroll-manager'
 import { SimpleEventEmitter } from './utils/simple-event-emitter'
 import { Autocapture } from './autocapture'
+import { TracingHeaders } from './extensions/tracing-headers'
 import { ConsentManager } from './consent'
 
 /*
@@ -173,6 +174,7 @@ export const defaultConfig = (): PostHogConfig => ({
     disable_compression: false,
     session_idle_timeout_seconds: 30 * 60, // 30 minutes
     person_profiles: 'always',
+    __add_tracing_headers: false,
 })
 
 export const configRenames = (origConfig: Partial<PostHogConfig>): Partial<PostHogConfig> => {
@@ -260,6 +262,7 @@ export class PostHog {
     analyticsDefaultEndpoint: string
 
     SentryIntegration: typeof SentryIntegration
+    sentryIntegration: (options?: SentryIntegrationOptions) => ReturnType<typeof sentryIntegration>
 
     private _debugEventEmitter = new SimpleEventEmitter()
 
@@ -271,8 +274,10 @@ export class PostHog {
 
     constructor() {
         this.config = defaultConfig()
+
         this.decideEndpointWasHit = false
         this.SentryIntegration = SentryIntegration
+        this.sentryIntegration = (options?: SentryIntegrationOptions) => sentryIntegration(this, options)
         this.__request_queue = []
         this.__loaded = false
         this.analyticsDefaultEndpoint = '/e/'
@@ -391,6 +396,8 @@ export class PostHog {
 
         this.sessionManager = new SessionIdManager(this.config, this.persistence)
         this.sessionPropsManager = new SessionPropsManager(this.sessionManager, this.persistence)
+
+        new TracingHeaders(this).startIfEnabledOrStop()
 
         this.sessionRecording = new SessionRecording(this)
         this.sessionRecording.startIfEnabledOrStop()
@@ -761,11 +768,12 @@ export class PostHog {
         // with every event and used by the session table to create session-initial props.
         if (this.config.store_google) {
             this.sessionPersistence.update_campaign_params()
-            this.persistence.set_initial_campaign_params()
         }
         if (this.config.save_referrer) {
             this.sessionPersistence.update_referrer_info()
-            this.persistence.set_initial_referrer_info()
+        }
+        if (this.config.store_google || this.config.save_referrer) {
+            this.persistence.set_initial_person_info()
         }
 
         let data: CaptureResult = {
@@ -1666,7 +1674,8 @@ export class PostHog {
      *         maskInputOptions: {},
      *         maskInputFn: null,
      *         slimDOMOptions: {},
-     *         collectFonts: false
+     *         collectFonts: false,
+     *         inlineStylesheet: true,
      *      }
      *
      *      // prevent autocapture from capturing any attribute names on elements