+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 1.139.1 - 2024-06-12
+
+- fix: less strict idle event dropping (#1241)
+- chore: add inlineStylesheet prop (#1237)
+
+## 1.139.0 - 2024-06-10
+
+- feat(personless-events): set_once more initial properties accurately when a person is created (#1236)
+
+## 1.138.3 - 2024-06-10
+
+- fix: circular refs shouldn't explode capture (#1230)
+
+## 1.138.2 - 2024-06-10
+
+- feat: Updated v8 sentry integration (#1224)
+
+## 1.138.1 - 2024-06-07
+
+- feat: wrap fetch and add tracing headers (#1186)
+- feat(survey): Support identifying iterations of a repeatable survey (#1200)
+
 ## 1.138.0 - 2024-06-06
 
 - feat(surveys): `posthog-js` changes to support valid HTML question rendering (#1206)